rename authenticator_state file
