+++ packages/amplify_authenticator/lib/src/state/inherited_authenticator_state.dart
